# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LoL Tracker API",
    "description": "Backend del dashboard de rendimiento en ranked. Sustituye al monolito Streamlit.",
    "version": "2.0.0-dev"
  },
  "paths": {
...
    "/api/sync": {
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Sync",
        "description": "Encola la sincronizaci\u00f3n y responde al instante (202).\n\nLa validaci\u00f3n (Riot ID, colas) sigue siendo s\u00edncrona: los errores de petici\u00f3n se reportan\naqu\u00ed, no en el polling. Los fallos de Riot/DB viajan despu\u00e9s por `/status`.",
        "operationId": "sync_api_sync_post",
        "parameters": [
          {
            "name": "riot_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Por defecto, el RIOT_ID del .env",
              "title": "Riot Id"
            },
            "description": "Por defecto, el RIOT_ID del .env"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "queues",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "IDs de cola separados por coma (420=Solo/Duo, 400=Normal Draft). Default: ambos.",
              "default": "420,400",
              "title": "Queues"
            },
            "description": "IDs de cola separados por coma (420=Solo/Duo, 400=Normal Draft). Default: ambos."
          },
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/sync/status": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Sync Status",
        "description": "Estado del sync para el polling del frontend (idle/processing/success/error).",
        "operationId": "sync_status_api_sync_status_get",
        "parameters": [
          {
            "name": "x-api-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LpCapture": {
        "properties": {
          "lp": {
            "type": "integer",
            "title": "Lp"
          },
          "tier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tier"
          },
          "division": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Division"
          },
          "delta_assigned": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delta Assigned"
          }
        },
        "type": "object",
        "required": [
          "lp"
        ],
        "title": "LpCapture",
        "description": "LP de Solo/Duo capturado autom\u00e1ticamente al cerrar el sync (League-V4).\n\n`delta_assigned` es el LP neto escrito como `lp_change` en la partida m\u00e1s reciente sin\nreview manual; None cuando no hab\u00eda snapshot previo (l\u00ednea base), delta 0, o ninguna\npartida candidata. Nunca sobrescribe una review del usuario."
      },
...
      "SyncAccepted": {
        "properties": {
          "status": {
            "type": "string",
            "const": "processing",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "SyncAccepted",
        "description": "Respuesta inmediata del POST /api/sync: el trabajo pesado corre en BackgroundTasks."
      },
      "SyncError": {
        "properties": {
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          }
        },
        "type": "object",
        "required": [
          "game_id",
          "reason",
          "retryable"
        ],
        "title": "SyncError"
      },
      "SyncResult": {
        "properties": {
          "fetched": {
            "type": "integer",
            "title": "Fetched"
          },
          "inserted": {
            "type": "integer",
            "title": "Inserted"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/SyncError"
            },
            "type": "array",
            "title": "Errors",
            "default": []
          },
          "lp_captured": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LpCapture"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "fetched",
          "inserted",
          "skipped"
        ],
        "title": "SyncResult",
        "description": "El sync ya no miente. El monolito dec\u00eda \"\u2728 Todo actualizado\" incluso cuando un 429 hab\u00eda\ndescartado partidas en silencio; ahora los fallos viajan en la respuesta."
      },
      "SyncStatus": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "idle",
              "processing",
              "success",
              "error"
            ],
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SyncResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "SyncStatus",
        "description": "Estado del sync en curso (o del \u00faltimo terminado), para el polling del frontend."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}